+++ src/tests/createHeadlessForm.test.js
@@ -1993,9 +1993,9 @@ describe('createHeadlessForm', () => {
             },
           },
           {
-            name: 'a_or_b',
-            label: 'A dropdown',
-            type: 'select',
+            name: 'role',
+            label: 'Role',
+            type: 'text',
             statement: {
               description: 'This is another statement message, but more severe.',
               inputType: 'statement',
@@ -2086,10 +2086,7 @@ describe('createHeadlessForm', () => {
       });
 
       expect(result).toMatchObject({
-        fields: [
-          { description: 'I am regular' },
-          { description: '<span class="jsf-description">I am <b>bold</b>.</span>' },
-        ],
+        fields: [{ description: 'I am regular' }, { description: 'I am <b>bold</b>.' }],
       });
     });
 
@@ -2548,7 +2545,7 @@ describe('createHeadlessForm', () => {
   });
 
   describe('when a field has conditional presentation properties', () => {
-    it('adds .jsf-statement to nested statement markup when visible', () => {
+    it('returns the nested properties when the conditional matches', () => {
       const { fields } = createHeadlessForm(schemaWithConditionalPresentationProperties, {
         initialValues: {
           // show the hidden statement
@@ -2556,9 +2553,7 @@ describe('createHeadlessForm', () => {
         },
       });
 
-      expect(fields[0].statement.description).toBe(
-        `<span class="jsf-statement"><a href="">conditional statement markup</a></span>`
-      );
+      expect(fields[0].statement.description).toBe(`<a href="">conditional statement markup</a>`);
     });
   });
 
@@ -3756,7 +3751,7 @@ describe('createHeadlessForm', () => {
       expect(fields).toMatchObject([
         {
           name: 'time',
-          description: '<span class="jsf-description">Write in <b>hh:ss</b> format</span>', // from presentation
+          description: 'Write in <b>hh:ss</b> format', // from presentation
           inputType: 'clock', // arbitrary type from presentation
           deprecated: {
             description: 'In favor of X', // from presentation